Create server in schedule instead of mocks
- Create express app with post route on port 3002 for immediateAdd
