+++ src/config/fieldNames.js
@@ -188,5 +188,5 @@ module.exports = {
         add: 'add',
         update: 'update',
     },
-    kartoffelUrl: "http://192.168.99.100:3000",
+    kartoffelUrl: "http://localhost:3000",
 };
