+++ src/util/data_synchronizeData.js
@@ -11,16 +11,7 @@ axios.defaults.headers.common['authorization'] = process.env.SOURCES_TOKEN;
 module.exports = async (dataSource, runnigType, ImmediateRunException = false) => {
     const dateAndTime = moment(new Date()).format("DD.MM.YYYY__HH.mm");
     let data;
-    // if(runnigType == fn.runnigTypes.ImmediateRun && !ImmediateRunException){
-    //     data = getPeopleFromServer();
-    //     await getPeopleFromServer().then(res => {
-    //         data = res;
-    //         // console.log(data);
-    //         dataSource = data[0];
-    //     });
-    //     // console.log(dataSource);
-    //     data.splice(0, 1);
-    // }
+
     if(runnigType == fn.runnigTypes.ImmediateRun && !ImmediateRunException){
         let receivedData = await getPeopleFromServer();
         data = receivedData.objects;
